+++ CHANGELOG.md
@@ -16,6 +16,17 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+## 2023-07-20
+
+- @triton-one/yellowstone-grpc:0.2.0
+- yellowstone-grpc-client-1.8.0+solana.1.16.1
+- yellowstone-grpc-geyser-1.5.0+solana.1.16.1
+- yellowstone-grpc-proto-1.8.0+solana.1.16.1
+
+### Features
+
+- geyser: add `Entry` message ([#163](https://github.com/rpcpool/yellowstone-grpc/pull/163)).
+
 ## 2023-07-18
 
 - yellowstone-grpc-geyser-1.4.0+solana.1.16.1